+++ src/main/java/city/TouristDrone.java
@@ -62,25 +62,4 @@ public class TouristDrone extends Drone
         File file = new File("/home/cybereagle3-1/AndroidStudioProjects/FirstApp/midsem_project/src/main/java/filestuff/tourists.txt");
         new FileIO().WriteUsersonFile(tourist, file);
     }
-
-    public void TouristMenu()
-    {
-        Scanner input = new Scanner(System.in);
-        boolean exit_loop = false;
-
-        while (!exit_loop)
-        {
-            System.out.println("welcome. what would you like to do");
-            System.out.println("1.\n2.cop login\n3.tourist admin login\n4.cop admin login\n5.shutdown system");
-            int choice = Integer.parseInt(input.nextLine());
-
-            if (choice == 1)
-            {
-            } else if (choice == 2)
-            {
-
-            }
-
-        }
-    }
 }