+++ src/main/java/users/Admin.java
@@ -23,28 +23,5 @@ public abstract class Admin extends User
         File read_file = new File("/home/cybereagle3-1/AndroidStudioProjects/FirstApp/midsem_project/src/main/java/filestuff/" + filename);
         new FileIO().DeleteUsersfromFile(x,read_file);
     }
-
-    /*public boolean AdminLogin(String admin_filename,String ID,String password) throws IOException
-    {
-        File file = new File("/home/cybereagle3-1/AndroidStudioProjects/FirstApp/midsem_project/src/main/java/filestuff/" + admin_filename);
-        BufferedReader file_reader = new BufferedReader(new FileReader(file));
-
-        String trimmed_line,current_line;
-        while((current_line = file_reader.readLine()) != null)
-        {
-            trimmed_line = current_line.trim();
-            if(trimmed_line.equals(ID))
-            {
-                current_line = file_reader.readLine();
-                trimmed_line = current_line.trim();
-
-                if (trimmed_line.equals(password))
-                    return true;
-            }
-        }
-
-        file_reader.close();
-        return false;
-    }*/
 }